Problem_Statement_3
Verilog module to implement a 1-to-2 demultiplexer by using gate level structural description. The module will take as arguments the following:
1-bit data input “in”,
1-bit select line “sel”
2-bit output “out”
module demux1to2 (out, in, sel);
  input in, sel;
  output [1:0] out;

Hint: If sel input signal is low, the input signal in will pass through the output line out[1]; otherwise if  sel input signal is high, the signal in will pass through the output line out[0]. The output line that is not selected must be at logic 0.